Update patch set 2
Patch Set 2:

easiest way would be to tag it to sensu-027-testing.

In my understanding, tagging it to sensu-common must be enough. However, since enabling multi-architectures for opstools, this seems to be broken, but neither me, nor centos infra folks could find out, what is exactly going on and how to fix that.

Easiest semi-fix would be to disable multi-architectures, but that would probably make ovirt folks quite unhappy, as they drove this.

Patch-set: 2